Split file generation logic apart because fs import breaks mobile
